+++ src/dungeon.h
@@ -14,8 +14,6 @@
 #include <exception.h>
 #include <global.h>
 
-#include "save_load.h"
-
 namespace App {
     class Dungeon {
 
@@ -40,6 +38,7 @@ namespace App {
         Settings* getSettings();
         EventManager* getEventManager();
         Player* getPlayer();
+        Monster* getBoss();
         WINDOW* getWindow();
         Output* getOutput();
         std::string getText(u_char index);
@@ -47,10 +46,13 @@ namespace App {
 
         /** SETTERS **/
         Dungeon* addFloor(Floor* floor, u_char index = U_CHAR_MAX);
+        Dungeon* removeMonsterTemplate(MonsterTemplate* monsterTemplate);
+        Dungeon* removeObjectTemplate(ObjectTemplate* objectTemplate);
         Dungeon* setCurrentFloor(Floor* floor);
         Dungeon* setSettings(Settings* settings);
         Dungeon* setEventManager(EventManager* eventManager);
         Dungeon* setPlayer(Player* player);
+        Dungeon* setBoss(Monster* boss);
         Dungeon* setWindow(WINDOW* window);
         /** SETTERS **/
     protected:
@@ -64,6 +66,7 @@ namespace App {
         Settings* settings;
         EventManager* eventManager;
         Player* player;
+        Monster* boss;
         Output* output;
 
         WINDOW* window;
@@ -72,6 +75,7 @@ namespace App {
     };
 }
 
+#include "save_load.h"
 #include "floor.h"
 #include "dice.h"
 #include "events.h"