+++ src/templates/MonsterTemplate.h
@@ -3,16 +3,15 @@
 
 #define MONSTER_TEMPLATE_HEADING "RLG327 MONSTER DESCRIPTION 1"
 
-#define MONSTER_TEMPLATE_NAME_VALID     0b000000001u
-#define MONSTER_TEMPLATE_DESC_VALID     0b000000010u
-#define MONSTER_TEMPLATE_COLOR_VALID    0b000000100u
-#define MONSTER_TEMPLATE_SPEED_VALID     0b000001000u
-#define MONSTER_TEMPLATE_ABILITIES_VALID     0b000010000u
-#define MONSTER_TEMPLATE_HIT_POINTS_VALID     0b000100000u
-#define MONSTER_TEMPLATE_ATTACK_DAMAGE_VALID     0b001000000u
-#define MONSTER_TEMPLATE_SYMBOL_VALID     0b010000000u
-#define MONSTER_TEMPLATE_RARITY_VALID     0b100000000u
-#define MONSTER_TEMPLATE_IS_VALID       0b111111111u
+#define MONSTER_TEMPLATE_NAME           "NAME"
+#define MONSTER_TEMPLATE_DESCRIPTION    "DESC"
+#define MONSTER_TEMPLATE_COLOR          "COLOR"
+#define MONSTER_TEMPLATE_SPEED          "SPEED"
+#define MONSTER_TEMPLATE_ABILITY        "ABIL"
+#define MONSTER_TEMPLATE_HIT_POINTS      "HP"
+#define MONSTER_TEMPLATE_DAMAGE         "DAM"
+#define MONSTER_TEMPLATE_SYMBOL         "SYMB"
+#define MONSTER_TEMPLATE_RARITY     "RRTY"
 
 #include <string>
 #include <vector>
@@ -34,7 +33,6 @@ namespace App {
         Monster* generateMonster(Floor* floor, u_char x, u_char y);
 
         /** GETTERS **/
-        bool isValid();
         std::string getName();
         std::string getDescription();
         u_short getColor();
@@ -47,13 +45,10 @@ namespace App {
         /** GETTERS **/
 
         /** SETTERS **/
-        MonsterTemplate* setIsValid(bool isValid);
         /** SETTERS **/
     protected:
 
     private:
-        bool validTemplate;
-
         std::string name;
         std::string description;
         u_short color;