+++ src/templates/ObjectTemplate.h
@@ -3,24 +3,24 @@
 
 #define OBJECT_TEMPLATE_HEADING "RLG327 OBJECT DESCRIPTION 1"
 
-#define OBJECT_TEMPLATE_NAME_VALID                  0b00000000000001u
-#define OBJECT_TEMPLATE_DESC_VALID                  0b00000000000010u
-#define OBJECT_TEMPLATE_TYPE_VALID                  0b00000000000100u
-#define OBJECT_TEMPLATE_COLOR_VALID                 0b00000000001000u
-#define OBJECT_TEMPLATE_HIT_BONUS_VALID             0b00000000010000u
-#define OBJECT_TEMPLATE_DAMAGE_BONUS_VALID          0b00000000100000u
-#define OBJECT_TEMPLATE_DODGE_BONUS_VALID           0b00000001000000u
-#define OBJECT_TEMPLATE_DEFENSE_BONUS_VALID         0b00000010000000u
-#define OBJECT_TEMPLATE_WEIGHT_VALID                0b00000100000000u
-#define OBJECT_TEMPLATE_SPEED_BONUS_VALID           0b00001000000000u
-#define OBJECT_TEMPLATE_SPECIAL_ATTRIBUTE_VALID     0b00010000000000u
-#define OBJECT_TEMPLATE_VALUE_VALID                 0b00100000000000u
-#define OBJECT_TEMPLATE_ARTIFACT_STATUS_VALID       0b01000000000000u
-#define OBJECT_TEMPLATE_RARITY_VALID                0b10000000000000u
-#define OBJECT_TEMPLATE_IS_VALID                    0b11111111111111u
+#define OBJECT_TEMPLATE_NAME                "NAME"
+#define OBJECT_TEMPLATE_DESCRIPTION         "DESC"
+#define OBJECT_TEMPLATE_TYPE                "TYPE"
+#define OBJECT_TEMPLATE_COLOR               "COLOR"
+#define OBJECT_TEMPLATE_HIT_BONUS           "HIT"
+#define OBJECT_TEMPLATE_DAMAGE_BONUS        "DAM"
+#define OBJECT_TEMPLATE_DODGE_BONUS         "DODGE"
+#define OBJECT_TEMPLATE_DEFENSE_BONUS       "DEF"
+#define OBJECT_TEMPLATE_WEIGHT              "WEIGHT"
+#define OBJECT_TEMPLATE_SPEED_BONUS         "SPEED"
+#define OBJECT_TEMPLATE_SPECIAL_ATTRIBUTE   "ATTR"
+#define OBJECT_TEMPLATE_VALUE               "VAL"
+#define OBJECT_TEMPLATE_ARTIFACT_STATUS     "ATR"
+#define OBJECT_TEMPLATE_RARITY              "RRTY"
 
 #include <string>
 #include <vector>
+#include <cstring>
 #include <iostream>
 #include <fstream>
 #include <sstream>
@@ -42,7 +42,6 @@ namespace App {
         Object* generateObject(Floor* floor, u_char x, u_char y);
 
         /** GETTERS **/
-        bool isValid();
         std::string getName();
         std::string getDescription();
         u_int getItemType();
@@ -60,13 +59,10 @@ namespace App {
         /** GETTERS **/
 
         /** SETTERS **/
-        ObjectTemplate* setIsValid(bool isValid);
         /** SETTERS **/
     protected:
 
     private:
-        bool validTemplate;
-
         std::string name;
         std::string description;
         u_int itemType;